+++ src/main/java/christmas/domain/DiscountEvents.java
@@ -1,4 +1,4 @@
-package christmas;
+package christmas.domain;
 
 import java.util.Arrays;
 import java.util.Map;
