+++ src/main/java/christmas/domain/Menu.java
@@ -1,4 +1,4 @@
-package christmas;
+package christmas.domain;
 
 import christmas.exception.ErrorMessage;
 import christmas.exception.InvalidOrderException;
