+++ src/main/java/christmas/domain/Type.java
@@ -1,4 +1,4 @@
-package christmas;
+package christmas.domain;
 
 public enum Type {
     Appetizer, MainCourse, Drinks, Desserts, Promotion;
