+++ src/main/java/christmas/view/InputView.java
@@ -1,4 +1,4 @@
-package christmas;
+package christmas.view;
 
 import camp.nextstep.edu.missionutils.Console;
